requirements.txt
asgiref==3.8.1
Django==4.2.17
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
load-dotenv==0.1.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
